feat/validator: refactor unit tests and modify log level of unformatted validation errors
